refactor(experiments): modularize network dataset generation pipeline
- Extract DAG configuration generation to separate module
- Extract naming variant generation to separate module
- Extract CPT variant generation to separate module
- Add dataset export functionality for HuggingFace-compatible parquet files
- Replace verbose print statements with tqdm progress bars
- Add missing dependencies (pandas, pyarrow, tqdm) to pyproject.toml
- Remove old notebook file in favor of modular Python scripts
- Add parquet files to gitignore
- Streamline main orchestration script (425 → 108 lines)

The new modular architecture enables:
- Better separation of responsibilities
- Clean progress tracking with progress bars
- Reproducible HuggingFace dataset generation
- Easy ablation testing on DAG structure, naming, and CPT parameters
